FIX: Correct pdf layout

+++ app/src/main/java/com/example/simpleexchangeprotocol/functionality/PictureHandler.java
@@ -70,7 +70,7 @@ public class PictureHandler {
 
     public static Bitmap getResizedBitmap(Bitmap bm, int newWidth){
         int width = bm.getWidth();
-        float scale = newWidth/width;
+        float scale = (float) newWidth/width;
         return  rescaleBitmap(bm, scale);
     }
